Add test for most of Pandoc and implement parts of it
This commit is too big.

The commit adds tests for most of the data types of Pandoc (with help from an
LLM). The plan is to use the tests to drive the development.

The commit separates inline and block, which are separate datatypes in the
haskell definition.

The commit adds support for most of the inline constructs and a few more block
constructs.