Use hickory to normalize html
Comparing the strings produced by Hiccup and Pandoc is a fools game. Most
pressing is the that the order of the HTML attributes have to occur in the same
order.

By parsing the strings back into a data structure (using hickory) we normalize
the data. This means that the produced HTML doesn't have to be *exactly* the
same. Examples of where this normalization comes in handy is that HTML
attributes does not need to occur in the same order, and whitespace is largely
ignored.