Remove hickory and use pandoc for normalization
When hickory parses html, the result is sensitive to whitespace (at least to
some degree). Hiccup does not use multiple lines when it is rendered to string.
This lead to us adding newlines to the hiccup, which seems like a bad idea.

By using pandoc to parse the pandoc back to JSON seems like a decent way of
normalizing the output, using the dependency we already are using (for
testing).