Allow ir->hiccup to transform chunks of pandocir
ir->hiccup assumed that the input was a full document, with :blocks being a
key. This is a bit inconvenient if you have extracted some subsets of the
blocks and would like to convert those into hiccup. This change retains the
previous behavior if :blocks is present, otherwise it performs the conversion
on the input directly.